Some error corrections
Make sure the tmp directory exists, or else the downloads won't work
right.

+++ Includes/AutoUpdate.php
@@ -84,14 +84,16 @@ Class AutoUpdate {
                 if( isset($data[0]['sha']) && $commitdata[0]['sha'] == $data[0]['sha']) {
                     $success = $this->pullcontents( 'https://api.github.com/repos/cyberpower678/Peachy/contents', false, $downloaddata, true );                
                 } elseif( isset($data[0]['sha']) && $commitdata[0]['sha'] != $data[0]['sha']) {
-                    unlink( $pgIP.'tmp/commit.tmp' );
-                    unlink( $pgIP.'tmp/download.tmp' );
+                    if( file_exists($pgIP.'tmp') ) mkdir($pgIP.'tmp', 775);
+                    if( file_exists($pgIP.'tmp/commit.tmp') ) unlink( $pgIP.'tmp/commit.tmp' );
+                    if( file_exists($pgIP.'tmp/download.tmp') ) unlink( $pgIP.'tmp/download.tmp' );
                     file_put_contents( $pgIP.'tmp/commit.tmp', serialize($data) );
                     $success = $this->pullcontents();                
                 } else {
                    $success = false; 
                 }
             } else {
+                if( file_exists($pgIP.'tmp') ) mkdir($pgIP.'tmp', 775);
                 file_put_contents( $pgIP.'tmp/commit.tmp', serialize($data) );
                 //$data = $this->processreturn( $data );
                 $success = $this->pullcontents();
@@ -99,8 +101,8 @@ Class AutoUpdate {
         }
         if( $success ) {
             file_put_contents( $pgIP.'Includes/Update.log', serialize($data) );
-            unlink( $pgIP.'tmp/commit.tmp' );
-            unlink( $pgIP.'tmp/download.tmp' );
+            if( file_exists($pgIP.'tmp/commit.tmp') ) unlink( $pgIP.'tmp/commit.tmp' );
+            if( file_exists($pgIP.'tmp/download.tmp') ) unlink( $pgIP.'tmp/download.tmp' );
             pecho( "Peachy Updated!  Changes will go into effect on the next run.\n\n", PECHO_NOTICE );
             return true;
         } else {
